[gRest] Cast timestamps to int (#1460)
* Cast all timestamps to integers

* Missed cast in epoch_info

+++ files/grest/rpc/pool/pool_updates.sql
@@ -1,7 +1,7 @@
 CREATE FUNCTION grest.pool_updates (_pool_bech32 text DEFAULT NULL)
     RETURNS TABLE (
         tx_hash text,
-        block_time numeric,
+        block_time integer,
         pool_id_bech32 character varying,
         pool_id_hex text,
         active_epoch_no bigint,
@@ -24,7 +24,7 @@ BEGIN
     RETURN QUERY
     SELECT
         tx_hash,
-        block_time,
+        block_time::integer,
         pool_id_bech32,
         pool_id_hex,
         active_epoch_no,
